Added documentation and some minor changes

+++ src/main/java/Crank_Bot/commands/ClearChatCommand.java
@@ -9,9 +9,9 @@ import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.MessageHistory;
 import net.dv8tion.jda.api.entities.TextChannel;
 
-public class ClearChat extends Command {
+public class ClearChatCommand extends Command {
 
-	public ClearChat() {
+	public ClearChatCommand() {
 		this.name = "clear-chat";
 		this.aliases = new String[] {"clean"};
 		this.help = "Cleans the chat of commands msgs";
@@ -19,15 +19,20 @@ public class ClearChat extends Command {
 	
 	@Override
 	protected void execute(CommandEvent event) {
+		
+		if (!event.getMember().isOwner()) {
+			event.reply("SoRry You aRe nOt perMitTed to PerfoRm suCh actioNs!");
+			return;
+		}
+		
 		TextChannel channel = event.getTextChannel();
 		event.reply("ClEarInG Chat cOmmAnDs & BOt mEssAgeS...");
 		
-		MessageHistory history = MessageHistory.getHistoryBefore(channel, event.getMessage().getId()).limit(50).complete();
+		MessageHistory history = MessageHistory.getHistoryBefore(channel, event.getMessage().getId()).complete();
 		List<Message> messages = history.getRetrievedHistory();
 		for (Message message : messages) {
 			boolean command = message.getContentRaw().matches(event.getClient().getPrefix() + ".*?");
 			boolean bot = message.getAuthor().isBot();
-			//event.reply(message.getContentRaw());
 			if (command || bot) {
 				message.delete().queue();
 			}
